requirements.txt
-i https://mirrors.aliyun.com/pypi/simple/
--trusted-host mirrors.aliyun.com

fastapi>=0.110.0
uvicorn>=0.27.0
pymysql>=1.1.0
sqlalchemy>=2.0
pydantic>=2.0
python-dotenv>=1.0
requests>=2.31.0